Move limit enforcing functions to utilities
